resource: refactor directory structure
Currently, the resource algebra directory has three different things:
- The RA/PCM traits
- RA combinators
- Types that hold RAs, meant for off-the-shelf usage

Without breaking clients (imports stay the same), this commit makes this
distinction in the directory structure:
- the traits/theory of RAs stays in the base dir
- the combinators get their own directory
- the implementations get their own directory
