+++ source/vstd/resource/combinators/frac.rs
@@ -1,13 +1,13 @@
-use super::super::modes::*;
-use super::super::prelude::*;
-use super::Loc;
-use super::agree::AgreementRA;
-use super::algebra::Resource;
-use super::algebra::ResourceAlgebra;
-use super::pcm::PCM;
-use super::product::ProductRA;
-use super::storage_protocol::*;
-use super::*;
+use super::super::super::modes::*;
+use super::super::super::prelude::*;
+use super::super::Loc;
+use super::super::agree::AgreementRA;
+use super::super::algebra::Resource;
+use super::super::algebra::ResourceAlgebra;
+use super::super::pcm::PCM;
+use super::super::product::ProductRA;
+use super::super::storage_protocol::*;
+use super::super::*;
 
 verus! {
