+++ source/vstd/resource/combinators/option.rs
@@ -1,7 +1,7 @@
-use super::super::prelude::*;
-use super::algebra::ResourceAlgebra;
-use super::pcm::PCM;
-use super::relations::*;
+use super::super::super::prelude::*;
+use super::super::algebra::ResourceAlgebra;
+use super::super::pcm::PCM;
+use super::super::relations::*;
 
 verus! {
 
@@ -90,7 +90,7 @@ pub proof fn lemma_set_op_opt<RA: ResourceAlgebra>(s: ISet<RA>, t: RA)
     ensures
         set_op(s, t).map(|b| Some(b)) == set_op(s.map(|x| Some(x)), Some(t)),
 {
-    broadcast use super::super::iset::group_iset_lemmas;
+    broadcast use super::super::super::iset::group_iset_lemmas;
 
     let s_mapped = s.map(|x| Some(x));
     let original = set_op(s, t);
