+++ source/vstd/resource/combinators/sum.rs
@@ -1,5 +1,5 @@
-use super::super::prelude::*;
-use super::algebra::ResourceAlgebra;
+use super::super::super::prelude::*;
+use super::super::algebra::ResourceAlgebra;
 
 verus! {
