+++ source/vstd/resource/impls/frac_opt.rs
@@ -1,15 +1,15 @@
-use super::super::modes::*;
-use super::super::prelude::*;
-use super::Loc;
-use super::storage_protocol::*;
+use super::super::super::modes::*;
+use super::super::super::prelude::*;
+use super::super::Loc;
+use super::super::storage_protocol::*;
 
 verus! {
 
 broadcast use {
-    super::super::imap::group_imap_lemmas,
-    super::super::iset::group_iset_lemmas,
-    super::super::map::group_map_lemmas,
-    super::super::set::group_set_lemmas,
+    super::super::super::imap::group_imap_lemmas,
+    super::super::super::iset::group_iset_lemmas,
+    super::super::super::map::group_map_lemmas,
+    super::super::super::set::group_set_lemmas,
 };
 
 /////// Fractional tokens that allow borrowing of resources
