+++ source/vstd/resource/impls/ghost_var.rs
@@ -1,8 +1,8 @@
 #[cfg(verus_keep_ghost)]
-use super::super::modes::tracked_swap;
-use super::super::prelude::*;
-use super::Loc;
-use super::frac::FracGhost;
+use super::super::super::modes::tracked_swap;
+use super::super::super::prelude::*;
+use super::super::Loc;
+use super::super::frac::FracGhost;
 
 verus! {
